refactor: 자동차 객체를 따로 만들어서 함수마다 전달하지 않고, 원하는 곳에서만 사용하도록 구현
