+++ src/game/startRacingGame.js
@@ -1,8 +1,9 @@
 import playCarRacing from './playCarRacing.js';
+import { cars } from '../init/cars.js';
 
-export default function startRacingGame(racingCount, carObjectList) {
+export default function startRacingGame(racingCount) {
   for (let i = 0; i < racingCount; i++) {
-    playCarRacing(carObjectList);
+    playCarRacing();
   }
-  console.log(carObjectList);
+  console.log(cars);
 }
